created new tools directory

+++ run.sh
@@ -33,7 +33,7 @@ run_single() {
 	while [ $i -lt $num ]
 	do
 		# remove existing output
-		rm output/*
+		rm -f output/*
 
 		repoName=${repositories[i]}
 		dirName="${repoName##*/}"
@@ -43,18 +43,34 @@ run_single() {
 
 		# Oxygen
 		python3 -m cli -a oxygen $repoName
+		if [ $? -ne 0 ]
+		then
+			echo "PyClone (Oxygen) - unable to process" >> $outputFile
+			((i=i+1))
+			cd ../..
+			continue
+		fi
+
 		jsonFiles=(*.json)
 		oxygenFile="${jsonFiles[0]}"
 		mv $oxygenFile $oxygenFilename
-		cp $oxygenFilename ../../output
+		cp $oxygenFilename ../../$TOOL_OUTPUT
 		rm $oxygenFilename
 
 		# Chlorine
 		python3 -m cli -a chlorine $repoName
+		if [ $? -ne 0 ]
+		then
+			echo "PyClone (Chlorine) - unable to process" >> $outputFile
+			((i=i+1))
+			cd ../..
+			continue
+		fi
+
 		jsonFiles=(*.json)
 		chlorineFile="${jsonFiles[0]}"
 		mv $chlorineFile $chlorineFilename
-		cp $chlorineFilename ../../output
+		cp $chlorineFilename ../../$TOOL_OUTPUT
 		rm $chlorineFilename
 
 		cd ..
@@ -64,27 +80,29 @@ run_single() {
 
 		# Blocks
 		./nicad5 blocks py ../../repos/$dirName
-		cp ../../repos/$dirName\_blocks-blind-clones/$dirName\_blocks-blind-clones-0.30.xml ../../output
+		cp ../../repos/$dirName\_blocks-blind-clones/$dirName\_blocks-blind-clones-0.30.xml ../../$TOOL_OUTPUT
 		# Stop if error
 		if [[ $? -ne 0 ]]
 		then
 			echo "NiCad Blocks - unable to process" >> $outputFile
 			((i=i+1))
+			cd ../..
 			continue
 		fi
-		mv ../../output/$dirName\_blocks-blind-clones-0.30.xml ../../output/$nicadBlocksFilename
+		mv ../../$TOOL_OUTPUT/$dirName\_blocks-blind-clones-0.30.xml ../../$TOOL_OUTPUT/$nicadBlocksFilename
 
 		# Functions
 		./nicad5 functions py ../../repos/$dirName
-		cp ../../repos/$dirName\_functions-blind-clones/$dirName\_functions-blind-clones-0.30.xml ../../output
+		cp ../../repos/$dirName\_functions-blind-clones/$dirName\_functions-blind-clones-0.30.xml ../../$TOOL_OUTPUT
 		# Stop if error
 		if [[ $? -ne 0 ]] 
 		then
 			echo "NiCad Functions - unable to process" >> $outputFile
 			((i=i+1))
+			cd ../..
 			continue
 		fi
-		mv ../../output/$dirName\_functions-blind-clones-0.30.xml ../../output/$nicadFunctionsFilename
+		mv ../../$TOOL_OUTPUT/$dirName\_functions-blind-clones-0.30.xml ../../$TOOL_OUTPUT/$nicadFunctionsFilename
 
 		cd ..
 
@@ -94,13 +112,13 @@ run_single() {
 		./flatten.sh ../../repos/$dirName
 		echo "Sending to Moss"
 		mossLink=$(./moss -l python ../../repos/$dirName/*.py | tail -n 1)
-		echo $mossLink > ../../output/$mossFilename
+		echo $mossLink > ../../$TOOL_OUTPUT/$mossFilename
 
 		cd ../..
 
 		# Process Results
 		echo "Processing Results"
-		stats=$(java -jar tools/clone-comparer/target/clone-comparer-1.0-SNAPSHOT.jar -M s -pO output/$oxygenFilename -pC output/$chlorineFilename -nB output/$nicadBlocksFilename -nF output/$nicadFunctionsFilename -m output/$mossFilename)
+		stats=$(java -jar tools/clone-comparer/target/clone-comparer-1.0-SNAPSHOT.jar -M s -pO $TOOL_OUTPUT/$oxygenFilename -pC $TOOL_OUTPUT/$chlorineFilename -nB $TOOL_OUTPUT/$nicadBlocksFilename -nF $TOOL_OUTPUT/$nicadFunctionsFilename -m $TOOL_OUTPUT/$mossFilename)
 		results="$dirName,$stats"
 		echo $results >> $outputFile
 
@@ -157,18 +175,34 @@ run_double() {
 
 			# Chlorine
 			python3 -m cli -a chlorine $repo1 $repo2
+			if [ $? -ne 0 ]
+			then
+				echo "PyClone (Chlorine) - unable to process" >> $outputFile
+				((j=j+1))
+				cd ../..
+				continue
+			fi
+
 			jsonFiles=(*.json)
 			chlorineFile="${jsonFiles[0]}"
 			mv $chlorineFile $chlorineFilename
-			cp $chlorineFilename ../../output
+			cp $chlorineFilename ../../$TOOL_OUTPUT
 			rm $chlorineFilename
 
 			# Iodine
 			python3 -m cli -a iodine $repo1 $repo2
+			if [ $? -ne 0 ]
+			then
+				echo "PyClone (Iodine) - unable to process" >> $outputFile
+				((j=j+1))
+				cd ../..
+				continue
+			fi
+
 			jsonFiles=(*.json)
 			iodineFile="${jsonFiles[0]}"
 			mv $iodineFile $iodineFilename
-			cp $iodineFilename ../../output
+			cp $iodineFilename ../../$TOOL_OUTPUT
 			rm $iodineFilename
 
 			cd ..
@@ -178,33 +212,35 @@ run_double() {
 
 			# Blocks
 			./nicad5cross blocks py ../../repos/$dir1 ../../repos/$dir2
-			cp ../../repos/$dir1\_blocks-blind-crossclones/$dir1\_blocks-blind-crossclones-0.30.xml ../../output
+			cp ../../repos/$dir1\_blocks-blind-crossclones/$dir1\_blocks-blind-crossclones-0.30.xml ../../$TOOL_OUTPUT
 			# Stop if error
 			if [[ $? -ne 0 ]]
 			then
 				echo "NiCad Blocks - unable to process" >> $outputFile
 				((j=j+1))
+				cd ../..
 				continue
 			fi
-			mv ../../output/$dir1\_blocks-blind-crossclones-0.30.xml ../../output/$nicadBlocksFilename
+			mv ../../$TOOL_OUTPUT/$dir1\_blocks-blind-crossclones-0.30.xml ../../$TOOL_OUTPUT/$nicadBlocksFilename
 
 			# Functions
 			./nicad5cross functions py ../../repos/$dir1 ../../repos/$dir2
-			cp ../../repos/$dir1\_functions-blind-crossclones/$dir1\_functions-blind-crossclones-0.30.xml ../../output
+			cp ../../repos/$dir1\_functions-blind-crossclones/$dir1\_functions-blind-crossclones-0.30.xml ../../$TOOL_OUTPUT
 			# Stop if error
 			if [[ $? -ne 0 ]] 
 			then
 				echo "NiCad Functions - unable to process" >> $outputFile
 				((j=j+1))
+				cd ../..
 				continue
 			fi
-			mv ../../output/$dir1\_functions-blind-crossclones-0.30.xml ../../output/$nicadFunctionsFilename
+			mv ../../$TOOL_OUTPUT/$dir1\_functions-blind-crossclones-0.30.xml ../../$TOOL_OUTPUT/$nicadFunctionsFilename
 
 			cd ../..
 
 			# Process Results
 			echo "Processing Results"
-			stats=$(java -jar tools/clone-comparer/target/clone-comparer-1.0-SNAPSHOT.jar -M d -pC output/$chlorineFilename -pI output/$iodineFilename -nB output/$nicadBlocksFilename -nF output/$nicadFunctionsFilename)
+			stats=$(java -jar tools/clone-comparer/target/clone-comparer-1.0-SNAPSHOT.jar -M d -pC $TOOL_OUTPUT/$chlorineFilename -pI $TOOL_OUTPUT/$iodineFilename -nB $TOOL_OUTPUT/$nicadBlocksFilename -nF $TOOL_OUTPUT/$nicadFunctionsFilename)
 			results="$dir1,$dir2,$stats"
 			echo $results >> $outputFile
 
@@ -217,6 +253,12 @@ run_double() {
 ###
 # Begin main script
 
+if [ "$1" == '' ] || [ "$1" == 'help' ]
+then
+	echo "Usage: <mode (single/double)> <repository URL file>"
+	exit 0
+fi
+
 # Handle arguments
 if [ $# -ne 2 ]
 then
@@ -227,21 +269,29 @@ fi
 mode=$1
 repository_file=$2
 
-if [ $mode != "single" ] && [ $mode != "double" ]
+if [ "$mode" != "single" ] && [ "$mode" != "double" ]
 then
 	echo "Mode must be single or double"
 	exit 2
 fi
 
+# Set output directories
+OUTPUT=output
+TOOL_OUTPUT=toolOutput
+
 # Build the Maven project
 cd tools/clone-comparer
-mvn clean package
-cd ../..
+mvn clean verify -q
+if [ $? -ne 0 ]
+then
+	exit $?
+fi
 
+cd ../..
 declare -a repositories
 
 # Declare output files
-outputFile=output-$mode-`date +%s`.csv
+outputFile=$OUTPUT/output-$mode-`date +%s`.csv
 declare -a tempFiles
 tempFiles[0]=oxygen.json
 tempFiles[1]=chlorine.json
@@ -259,7 +309,7 @@ do
 done < $repository_file
 
 # Iterate over the repositories
-mkdir repos
+mkdir -p repos
 cd repos
 i=0
 while [ $i -lt $num ]
@@ -270,16 +320,17 @@ done
 cd ..
 
 # Clear out any existing JSON files in PyClone
-rm tools/codeDuplicationParser/*.json
+rm -f tools/codeDuplicationParser/*.json
 
-#virtualenv=venv
-virtualenv=venv_pypy
+virtualenv=venv
+#virtualenv=venv_pypy
 
 # Activate Python virtual environment
 source tools/codeDuplicationParser/$virtualenv/bin/activate
 
-# Create output directory
-mkdir output
+# Create output directories
+mkdir -p $OUTPUT
+mkdir -p $TOOL_OUTPUT
 
 # Run the clone tools
 if [ $mode == 'single' ]
